api/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
psycopg2-binary>=2.9.9
drf-spectacular>=0.27.0
django-cors-headers>=4.3.0
Pillow>=10.0.0
python-dotenv>=1.0.0
